+++ lib/iden3comm/domain/repositories/iden3comm_repository.dart
@@ -21,7 +21,7 @@ abstract class Iden3commRepository {
   // FIXME: use Entities only from the same part
   Future<Uint8List> getAuthInputs(
       {required String genesisDid,
-      required int profileNonce,
+      required BigInt profileNonce,
       required String challenge,
       required List<String> authClaim,
       required IdentityEntity identity,
